requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
anthropic>=0.19.1
supabase==2.1.0
langgraph==0.0.30
pydantic==2.4.2
typing-extensions==4.8.0
asyncio==3.4.3
pytest==7.4.2
httpx>=0.24.0,<0.25.0 